Proof of concept refactor to allow non 'run' jobs
