+++ girder_worker/__init__.py
@@ -345,3 +345,57 @@ def run(task, inputs=None, outputs=None, auto_convert=True, validate=True,
         return outputs
     finally:
         girder_worker.events.trigger('run.finally', info)
+
+
+
+
+
+
+
+
+from .format import conv_graph
+from celery import Celery
+
+def task_includes():
+    return ['girder_worker']
+
+
+app = Celery(
+    main=girder_worker.config.get('celery', 'app_main'),
+    backend=girder_worker.config.get('celery', 'broker'),
+    broker=girder_worker.config.get('celery', 'broker')
+)
+
+
+@app.task(name='girder_worker.run')
+def run_task(*pargs, **kwargs):
+    jobInfo = kwargs.pop('jobInfo', {})
+    retval = 0
+
+    with utils.JobManager(logPrint=jobInfo.get('logPrint', True),
+                          url=jobInfo.get('url'), method=jobInfo.get('method'),
+                          headers=jobInfo.get('headers'),
+                          reference=jobInfo.get('reference')) as jm:
+        kwargs['_job_manager'] = jm
+        kwargs['status'] = utils.JobStatus.RUNNING
+        retval = girder_worker.run(*pargs, **kwargs)
+    return retval
+
+@app.task
+def convert(*pargs, **kwargs):
+    return girder_worker.convert(*pargs, **kwargs)
+
+
+@app.task
+def validators(*pargs, **kwargs):
+    _type, _format = pargs
+    nodes = []
+
+    for (node, data) in conv_graph.nodes(data=True):
+        if ((_type is None) or (_type == node.type)) and \
+           ((_format is None) or (_format == node.format)):
+            nodes.append({'type': node.type,
+                          'format': node.format,
+                          'validator': data})
+
+    return nodes
